requirements.txt
python-telegram-bot==20.7
openai==1.12.0
newspaper3k==0.2.8
urllib3==2.2.1
python-dotenv==1.0.0
lxml[html_clean]>=5.1.0
requests==2.31.0
beautifulsoup4==4.12.2 